Generalize layout selector component; add selector to Compact view; #103
